게시판 초안 작성 - 도메인 클래스 타입 수정  / 22-02-21 / 0.0.12v

+++ src/main/java/com/smartFarm/was/domain/model/ActivityLog.java
@@ -2,8 +2,8 @@ package com.smartFarm.was.domain.model;
 
 public class ActivityLog {
 
-    private Integer logId;
-    private Integer deviceAuto;
+    private long logId;
+    private long deviceAuto;
     private String transactionId;
     private String logContent;
     private String logAccessIp;
