+++ src/main/java/com/smartFarm/was/domain/model/DataType.java
@@ -2,7 +2,7 @@ package com.smartFarm.was.domain.model;
 
 public class DataType {
 
-    private Integer deviceOIId;
-    private Integer deviceId;
-    private Integer deviceOIType;
+    private long deviceOIId;
+    private long deviceId;
+    private long deviceOIType;
 }
