+++ src/main/java/com/smartFarm/was/domain/model/Device.java
@@ -2,7 +2,7 @@ package com.smartFarm.was.domain.model;
 
 public class Device {
 
-    private Integer deviceId;
+    private long deviceId;
     private String deviceAlias;
     private String deviceType;
     private String status;
